refactor(example): layer React Aria reference handlers (CA-REACT-ARIA-REF-001)
